+++ tests/test_crawler.py
@@ -12,6 +12,11 @@ URLS = ("https://a.co/d/geQhA2O",
         "https://a.co/d/3h52kBL")
 
 
-def test_crawler():
-    data = asyncio.run(crawler.scrap_urls(URLS))
+def test_crawler_lxml():
+    data = asyncio.run(crawler.scrap_urls(URLS, method="lxml"))
+    console.print(data)
+
+
+def test_crawler_playwright():
+    data = asyncio.run(crawler.scrap_urls(URLS, method="playwright"))
     console.print(data)